+++ 2021/02/main.py
@@ -1,17 +1,16 @@
-#!/usr/bin/env python3
-
-
 def read_file(path):
     with open(path, "r") as f:
         lines = []
         for l in f.readlines():
             line = l.strip().split()
-            lines.append((line[0], float(line[1])))
+            lines.append((line[0], int(line[1])))
 
         return lines
 
+
 def final_position(commands, use_aim):
     horizontal, depth, aim = 0, 0, 0
+
     for d, x in commands:
         if d == "forward":
             horizontal += x
@@ -26,6 +25,7 @@ def final_position(commands, use_aim):
 
     return horizontal * depth
 
+
 def main():
     commands = read_file("input.txt")
 
@@ -34,6 +34,7 @@ def main():
 
     print(f"Part 1: {solution_1}\n"
           f"Part 2: {solution_2}")
+
 
 if __name__ == '__main__':
     main()
